requirements.txt
torch>=2.1
torchvision>=0.16
numpy>=1.23
pandas>=1.5
pyyaml>=6.0
matplotlib>=3.7
scikit-learn>=1.2
tqdm>=4.66
